Fixes some issues that are reported.
 #37 #47 #48 #49 #50 #51 #80 #79 #75 #70 #105 #104 #103 #102 #101 #100 #96 #94 #92 #89 #82 #128 #125 #124 #123 #119 #113 #108 #155 #154 #153 #151 #148 #145 #143 #140 #137 #136 #181 #180 #179 #178 #177 #176 #175 #174 #173 #171 #169 #167 #160 #158 #157 #207 #206 #205 #203 #202 #201 #194 #193 #192 #188 #184 #186 #183 #182 #231 #230 #228 #227 #225 #223 #222 #220 #219 #218 #217 #216 #215 #214 #213 #212 #211 #210 #208

+++ KenshiModTool/MainWindow.xaml.cs
@@ -494,20 +494,30 @@ namespace KenshiModTool
 
         public void SaveModList_Click(object sender, RoutedEventArgs e)
         {
+
             if (File.Exists(Path.Combine(LoadService.config.GamePath, "data", "mods.cfg")))
                 File.Copy(Path.Combine(LoadService.config.GamePath, "data", "mods.cfg"), Path.Combine(LoadService.config.GamePath, "data", "mods.cfg.backup"), true);
 
-            File.WriteAllLines(Path.Combine(LoadService.config.GamePath, "data", "mods.cfg"), ModList.Where(m => m.Active).OrderBy(q => q.Order).Select(m => m.FileName));
+
+            if (Directory.Exists(Path.Combine(LoadService.config.GamePath)))
+            {
+                File.WriteAllLines(Path.Combine(LoadService.config.GamePath, "data", "mods.cfg"), ModList.Where(m => m.Active).OrderBy(q => q.Order).Select(m => m.FileName));
+            }
+            else
+            {
+                MessageBox.Show(ConstantMessages.GameFolderNotConfiguredCorrectly, "List not saved!");
+            }
+
         }
 
         public void BtnGameFolder_Click(object sender, RoutedEventArgs e)
         {
-            CommonService.OpenFolder(LoadService.config.GamePath, () => MessageBox.Show("Game folder not configured correctly."));
+            CommonService.OpenFolder(LoadService.config.GamePath, () => MessageBox.Show(ConstantMessages.GameFolderNotConfiguredCorrectly));
         }
 
         public void GameModFolder_Click(object sender, RoutedEventArgs e)
         {
-            CommonService.OpenFolder(System.IO.Path.Combine(LoadService.config.GamePath, "Mods"), () => MessageBox.Show("Game folder not configured correctly."));
+            CommonService.OpenFolder(System.IO.Path.Combine(LoadService.config.GamePath, "Mods"), () => MessageBox.Show(ConstantMessages.GameFolderNotConfiguredCorrectly));
         }
 
         public void BtnSteamFolder_Click(object sender, RoutedEventArgs e)
@@ -721,7 +731,7 @@ namespace KenshiModTool
 
             if (failure)
             {
-                MessageBox.Show("Game folder not configured correctly.");
+                MessageBox.Show(ConstantMessages.GameFolderNotConfiguredCorrectly);
             }
         }
